fix review findings
