+++ tests/cmake.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 import pytest
 
-from . import adb
+from . import adb, lib_name
 from .conditions import has_sccache
 from .build_config import (
     get_android_config,
@@ -75,13 +75,6 @@ class CMake:
 
             def exe_name(name):
                 return name + ".exe" if sys.platform == "win32" else name
-
-            def lib_name(name):
-                if sys.platform == "win32":
-                    return name + ".dll"
-                elif sys.platform == "darwin":
-                    return "lib" + name + ".dylib"
-                return "lib" + name + ".so"
 
             for i, (d, _) in enumerate(self.runs.values()):
                 # first merge the raw profiling runs
